requirements.txt
Flask==3.0.0
tensorflow==2.14.0
mediapipe==0.10.14
numpy==1.24.4
opencv-python==4.10.0.84
Pillow==9.2.0
gunicorn==20.1.0